Add an aggregator to transform game events into state
There are several views of game scores that are useful: the current
totals and the round-by-round scores. This aggregator calculates both,
based on the assumption that if rounds are important all players will
record scores for one round before moving on to the next round.

Like elsewhere, his uses classes internally but hides the
implementation, only exposing a factory function. This limits the places
that "new" needs to be called, making it easier to move to a fully
functional style later if that seems appropriate. Classes are used here
for managing internal state while state transformation logic has
deliberately been kept pure and separate from the class so it could be
refactored away from the internal state later, possibly (clearly
influenced by redux).