Rename domain & remove object spread
There was a problem importing the domain into the react app which
turned out to be because of object destructuring (not supported in the
import for some reason ...) so this has been replaced (it was taking
too long to fix the module loading).

Also, the domain model has been renamed to be more specific and so that
the package name matches the directory name, for ease.