Allow differences in spelling when looking up players by name
Alexa sometimes interprets names with different spellings from those
in a game (eg "Danyal" or "Jon" rather than "Daniel" or "John"). To
deal with this, I've used a heuristic to try and find names that were
probably intended. This uses the Levenshtein distance of the possible
name from all names in the game (the Levenshtein distance is the
number of single characters needed to be added, switched or removed
to make two strings equal). Using this distance divided by the length
of the name give a rough ratio of how different the strings are,
allowing an estimate to be made and clearly incorrect ones to be
filtered out.