Enable compiling web front end
This is the first step to creating a releasable server with both the
Alexa endpoints and serving the React app. It's a slightly tricky
process as the domain module has not been pubished, so needs to be
copied into the app and not linted (some of its dependencies don't
pass).