Add environment-specific config
Certian pieces of information depend on the environment the app is
running in. This is now abstracted into a single directory with an
index file that does the selection. The root URL needs to be specified
because it doesn't seem to be easily possible within an Alexa intent
handler to get the URL of the site.