Persist game events in MongoDB
Heroku sets up the MONGODB_URI environment variable to enable this
connection. Unfortunately, at the moment I don't have the bandwidth to
install Mongo locally, so am using a different collection on the remote
db provided by Heroku. If this moves beyond a hobby project, this will
need to be fixed.

Using a repository pattern means that persistence is easy to mock out,
so in dev, for example, I could easily go back to an in-memory
solution. It makes sense that the websocket connection ids for each
game are still only stored in memory as they only are useful to the
currently running server.

Also marks lock files as binary to stop them showing in diffs (I should
not be modifying these manually).