Persist events locally in the web ui
This enables the app to be used offline which is an important use case
for this - often board games and card games are played because there is
limited access to the internet.

This uses a localforage, a library that wraps IndexedDB, for local JS
persistence. It has a simple API and falls back to localStorage.
Similar to elsewhere, the domain and this library are passed in so that
they are easier to mock out that direct imports.

This also limits what is stored to just the events necessary to rebuild
essential state - game events and only one join per client/game pair.
This prevents too much being stored and prevents storing another event
on every connected client any time anyone reloads a page.

In combination with the earlier server connection management, this
means that any events generated while the user is offline will be
synced whenever they generate another event after coming back online.